got android emulator working
mostly self-explanatory, `nix run .#create-emulator` to make the right
emulator, then `nix run .#run-emulator` to run it. then you can `yarn
android` to build and run the android version. sometimes it seems to
crash on first startup, click through to expo and connect to the dev
server again in expo.